add ndctl to supported tools
This is part of the nvdimm efforts around pmdk and libpmem1.
libpmem1 will be pulled in by qemu, but ndctl is a tool and no
runtime dependency, therefore let us add that to the seed directly.

Adding src:ndctl will as usual also add all the -dev packages but
on ndctl that does not seem to pull in any other universe packages.
Therefore this should be enough, no extra-excludes needed.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>