rescue libaio into supported-hardware-server
libaio is only directly seeded (obviously many indirections are present)
in the to be removed server-ship seed. Add it to the list of
storage options and backends that is in supported-hardware-server.

libaio-dev was in server-ship as well which a) made no sense on the iso
and b) will be auto-promotes via the default Extra-Include rules.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>