autodetect version suffix on seconday postgresql binaries
postgresql creates several binaries that aren't seeded directly.
But the current dependencies use versioned suffixes like -10 which
would need to get updated and they were not. Due to that they got
auto-demoted after Bionic.

Fix that by using wildcards. Please do note that wildcards are not
documented in germinate, but used in other places e.g.:
  desktop-common:92: * !linux-image-*

A test run with germinate using package in main worked fine. The three
postgresql packages show up as:
  ? Unknown supported-misc-servers package: postgresql-plperl-*
  ? Unknown supported-misc-servers package: postgresql-plpython3-*
  ? Unknown supported-misc-servers package: postgresql-pltcl-*
Which is the component mismatch, when running with --components=main,universe
they are ok. We want this mismatch to promote them so this is correct.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>