Rescue ipmi tools from server-ship to supported-sysadmin-server
Only openipmi is from server-ship, but the manufacturers can't settle
on just one ipmi tool for all so also add freeipmi-tools which is in main
and only held by supported-maas to the same place.

ipmitool is the third of the group and also only held by community-maas but
currently not in main. It is WIP (LP: #1576812) but not currently part of
the server teams packages.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>