move openstack into a separate supported-openstack file
There are a few special cases which has use-cases in both worlds
(openstack and general server) like radvd and openvswitch.
Those are also co-maintained by both Teams right now.

Due to this we end up with a misc-server seed retaining those the
server team is interested in while the openstack seed shows what they
need.

Due to that there are some duplicates which we usually avoid, but those
other would be e.g. "misc-server + lamp" which both are server-team.
In this case it reflects that e.g. openvswitch is of interest for both
teams and their user-base.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>