server: move lsb-invalid-mta hint
supported-sysadmin-server is evaluated before supported-misc-servers and
contains checksecurity which will pull in logcheck.
That will depend on default-mta which will fail to resolve and make it
select esmtp-run (universe).

Move the lsb-invalid-mta hint from supported-misc-servers to be before
that avoiding the mismatch (which btw is i386 only as postfix isn't
there).

This might eventually be a bug in gemrinate (LP: #1895011) but until
then move this to avoid the false positive.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>