seed probert (needed by subiquity)
probert is a dependency of subiquity and should therefore be seeded. The
subiquity snap includes its own copy of probert, so things are not
broken, but we still want all the installer components to be in main for
consistency.

At the moment probert is already in main and not listed as "move to
universe" component-mismatch because probert-network does not exist for
i386, and curtin has a dependency on:

  probert-storage | probert

which keeps probert in main.

This change will make probert-network appear as a component-mismatch as
it is currently not in main. This is correct and will have to be fixed
via a MIR. This will fix LP: #1890301.