directly seed xz-utils in standard.
Was previously pulled in as a dependency of mime-support, and this is a
widely-used compression format that we should keep installed by default
along with gzip.  bzip2 has also dropped out of the seed, but not seeding
this directly as that is a compression format that is largely superseded
by xz in practice.