Drop php-xmlrpc from Main
This used to be carried in src:php and thus was in main, and as such has
been seeded in supported-misc-servers.  However, nothing in main
actually depends on it (only src:libsoup2.4 appears to, and that's a
universe package), and it's not a package we generally have been
providing direct end-user support for, so seems best suited to live in
universe now.

LP: #1956345
Signed-off-by: Bryce Harrington <bryce@bryceharrington.org>