Make both iptables and nftables optional
Since nftables is the successor to iptables it is likely users will want to
be able to uninstall iptables and just use nftables. However, there will
also likely be users who are happy with iptables and don't wish to move to
nftables yet. So make both be recommends so either can be uninstalled.

Signed-off-by: Alex Murray <alex.murray@canonical.com>