Unseed acpi-support
The package still includes a small number of ACPI events, each of which
appear to be handled elsewhere. Specifically, each remaining event either:

(1) Is covered in an appropriate hid- kernel driver, e.g. hid-asus
(2) Is covered in systemd's hwdb.d/60-keyboard.hwdb
(3) Matches an event string that was generated by the old
    /proc/acpi/event interface, and is therefore obsolete.